refactor: moved other miscellaneous objects to model folder
