+++ src/main/java/com/cewar/controllers/CardController.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.cewar.library.Card;
+import com.cewar.model.entity.Card;
 import com.cewar.repositories.CardRepository;
 import com.cewar.util.CardReader;
