+++ src/main/java/com/cewar/controllers/GameController.java
@@ -5,8 +5,8 @@ import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.util.HtmlUtils;
 
-import com.cewar.game.GameMessage;
-import com.cewar.game.GameResponse;
+import com.cewar.model.dtos.GameMessageDto;
+import com.cewar.model.dtos.GameResponseDto;
 
 /**
  * Controller for game API functions
@@ -18,7 +18,7 @@ public class GameController {
 
     @MessageMapping("/message")   // If message is sent to /message, call this method
     @SendTo("/game/update") // Broadcast response to all subscribers of /game/update
-    public GameResponse handleAction(GameMessage message) throws Exception {
+    public GameResponseDto handleAction(GameMessageDto message) throws Exception {
 
 
 
@@ -27,6 +27,6 @@ public class GameController {
 
 
         Thread.sleep(1000); // simulated delay
-        return new GameResponse("Received message " + HtmlUtils.htmlEscape(message.getMessage())); // Sanitize to prevent attacks
+        return new GameResponseDto("Received message " + HtmlUtils.htmlEscape(message.getMessage())); // Sanitize to prevent attacks
     }
 }
