+++ src/main/java/com/cewar/model/entity/User.java
@@ -1,4 +1,4 @@
-package com.cewar.web.userdetails;
+package com.cewar.model.entity;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -9,7 +9,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import com.cewar.enums.Authority;
-import com.cewar.library.Card;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
