+++ src/main/java/com/cewar/repositories/CardRepository.java
@@ -3,7 +3,7 @@ package com.cewar.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.cewar.library.Card;
+import com.cewar.model.entity.Card;
 
 /**
  * Repository to get card information from a database
