+++ src/main/java/com/cewar/util/CardReader.java
@@ -4,7 +4,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
-import com.cewar.library.Card;
+import com.cewar.model.entity.Card;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
